Moved Main into MemberMain to match #94
